day-task/day-6/Multi-agent-system/pyproject.toml
[project]
name = "agent-with-chat-historty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "dotenv>=0.9.9",
    "openai-agents>=0.0.19",
]
